+++ src/conn.c
@@ -1,5 +1,5 @@
 // SPDX-License-Identifier: BSD-3-Clause
-/* Copyright 2020-2021, Intel Corporation */
+/* Copyright 2020-2022, Intel Corporation */
 /* Copyright 2021, Fujitsu */
 
 /*
@@ -381,6 +381,26 @@ rpma_write_atomic(struct rpma_conn *conn,
 			RPMA_ATOMIC_WRITE_ALIGNMENT, flags,
 			IBV_WR_RDMA_WRITE, 0,
 			op_context, true);
+}
+
+/*
+ * rpma_atomic_write -- initiate the atomic 8 bytes write operation
+ */
+int
+rpma_atomic_write(struct rpma_conn *conn,
+	struct rpma_mr_remote *dst, size_t dst_offset,
+	const char src[8],
+	int flags, const void *op_context)
+{
+	if (conn == NULL || dst == NULL || src == NULL || flags == 0)
+		return RPMA_E_INVAL;
+
+	if (dst_offset % RPMA_ATOMIC_WRITE_ALIGNMENT != 0)
+		return RPMA_E_INVAL;
+
+	return rpma_mr_atomic_write(conn->id->qp,
+			dst, dst_offset, src,
+			flags, op_context);
 }
 
 /*