+++ src/mr.h
@@ -1,5 +1,5 @@
 /* SPDX-License-Identifier: BSD-3-Clause */
-/* Copyright 2020-2021, Intel Corporation */
+/* Copyright 2020-2022, Intel Corporation */
 
 /*
  * mr.h -- librpma memory region-related internal definitions
@@ -37,7 +37,7 @@ int rpma_mr_read(struct ibv_qp *qp,
  *    dst_offset == 0 && src_offset == 0 && len == 0)
  *
  * ERRORS
- * rpma_mr_write() can fail with the following error:
+ * rpma_mr_write() can fail with the following errors:
  *
  * - RPMA_E_NOSUPP   - unsupported 'operation' argument
  * - RPMA_E_PROVIDER - ibv_post_send(3) failed
@@ -47,6 +47,19 @@ int rpma_mr_write(struct ibv_qp *qp,
 	const struct rpma_mr_local *src,  size_t src_offset,
 	size_t len, int flags, enum ibv_wr_opcode operation,
 	uint32_t imm, const void *op_context, bool fence);
+
+/*
+ * ASSUMPTIONS
+ * - qp != NULL && dst != NULL && src != NULL && flags != 0
+ *
+ * ERRORS
+ * rpma_mr_atomic_write() can fail with the following error:
+ *
+ * - RPMA_E_PROVIDER - ibv_post_send(3) failed
+ */
+int rpma_mr_atomic_write(struct ibv_qp *qp,
+	struct rpma_mr_remote *dst, size_t dst_offset,
+	const char src[8], int flags, const void *op_context);
 
 /*
  * ASSUMPTIONS