+++ tests/unit/common/mocks-rpma-mr.c
@@ -1,5 +1,5 @@
 // SPDX-License-Identifier: BSD-3-Clause
-/* Copyright 2020-2021, Intel Corporation */
+/* Copyright 2020-2022, Intel Corporation */
 
 /*
  * mocks-rpma-mr.c -- librpma mr.c module mocks
@@ -71,11 +71,34 @@ rpma_mr_write(struct ibv_qp *qp,
 }
 
 /*
+ * rpma_mr_atomic_write -- rpma_mr_atomic_write() mock
+ */
+int
+rpma_mr_atomic_write(struct ibv_qp *qp,
+	struct rpma_mr_remote *dst, size_t dst_offset,
+	const char src[8], int flags, const void *op_context)
+{
+	assert_non_null(qp);
+	assert_int_not_equal(flags, 0);
+	assert_non_null(src);
+	assert_non_null(dst);
+
+	check_expected_ptr(qp);
+	check_expected_ptr(dst);
+	check_expected(dst_offset);
+	check_expected_ptr(src);
+	check_expected(flags);
+	check_expected_ptr(op_context);
+
+	return mock_type(int);
+}
+
+/*
  * rpma_mr_reg -- a mock of rpma_mr_reg()
  */
 int
 rpma_mr_reg(struct rpma_peer *peer, void *ptr, size_t size, int usage,
-		struct rpma_mr_local **mr_ptr)
+	struct rpma_mr_local **mr_ptr)
 {
 	check_expected_ptr(peer);
 	check_expected(size);